requirements.txt
requests==2.28.2
beautifulsoup4==4.11.2
pyyaml==6.0
readability-lxml==0.8.1
openai==0.27.8
anthropic==0.3.6